remove blog miner
This just isn't ready. Some adblockers say it's malware for some reason,
there's issues with connecting most of the time, and it's just not doing
much anyway. It's being removed until I can take the time to solve the
multitude of problems and/or re-evaluate if it's even worth it.